pyproject.toml
[project]
name = "colectivos-ws-formatter-insite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.25.0",
    "certifi>=2025.1.31",
    "deltalake>=0.25.4",
    "isodate>=0.7.2",
    "polars>=1.24.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
